I didn't realize there was an existing "customscan.rs" test, so merge everything together into the new one with an underscore
