--- pg_search/tests/customscan.rs
@@ -1,33 +1,0 @@
-// Copyright (c) 2023-2024 Retake, Inc.
-//
-// This file is part of ParadeDB - Postgres for Search and Analytics
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-mod fixtures;
-
-use fixtures::*;
-use pretty_assertions::assert_eq;
-use rstest::*;
-use shared::fixtures::db::Query;
-use sqlx::PgConnection;
-
-#[rstest]
-fn attribute_1_of_table_has_wrong_type(mut conn: PgConnection) {
-    SimpleProductsTable::setup().execute(&mut conn);
-
-    let (id,) = "SELECT id, description FROM paradedb.bm25_search WHERE description @@@ 'keyboard' OR id = 1 ORDER BY id LIMIT 1"
-        .fetch_one::<(i32,)>(&mut conn);
-    assert_eq!(id, 1);
-}
